work on 42 mac, can be pushed. domain cannot be changed from /etc/hosts because of root privileges

+++ srcs/wordpress/start.sh
@@ -1,10 +1,14 @@
-echo "\n[ii] downloading wordpress\n"
-wget https://wordpress.org/latest.tar.gz >/dev/null 2>/dev/null
-echo "\n[ii] wordpress has been downloaded\n"
-tar -xzf latest.tar.gz
-rm -rf /var/www/gcefalo/wordpress
-mv /wordpress /var/www/gcefalo/wordpress
-rm -rf latest.tar.gz
+
+if [ ! -d /var/www/gcefalo/wordpress ]; then
+    echo "\n- downloading wordpress\n"
+    wget https://wordpress.org/latest.tar.gz >/dev/null 2>/dev/null
+    echo "\n- wordpress has been downloaded\n"
+    tar -xzf latest.tar.gz
+    mv /wordpress /var/www/gcefalo/wordpress
+    rm -rf latest.tar.gz
+else
+    echo "\n- skipping wordpress download: wordpress already installed\n"
+fi
 
 mv /wp-config.php	/var/www/gcefalo/wordpress
 mv /php-fpm.conf	/etc/php/7.3/fpm/php-fpm.conf
